+++ src/components/AuthForm.js
@@ -1,15 +1,18 @@
-import React from "react";
-import useAuthContext from "../hooks/useAuthContext";
+import React, { useState } from "react";
 import LoginForm from "./LoginForm";
 import SignupForm from "./SignupForm";
 
 const AuthForm = () => {
-    const { userAction } = useAuthContext();
+    const [userAction, setUserAction] = useState("");
 
     return (
         <>
-            {userAction === "login" && <LoginForm />}
-            {userAction === "signup" && <SignupForm />}
+            {userAction === "login" && (
+                <LoginForm setUserAction={setUserAction} />
+            )}
+            {userAction === "signup" && (
+                <SignupForm setUserAction={setUserAction} />
+            )}
         </>
     );
 };
