+++ src/pages/Gamepage.css
@@ -10,9 +10,7 @@ body {
 
 h1,
 h2,
-h3,
-h5,
-h6 {
+h3 {
   font-family: "Artifika", serif;
   color: #efefef;
 }
@@ -22,6 +20,9 @@ h2 {
   line-height: 22px;
   font-weight: 100;
 }
+
+
+
 
 h3 {
   font-size: 16px;
